Add support to toggle LaTex rendering from editor
If focus is given to any editor view that has an associated markdown
preview, then toggling through the menu item or the ctrl-shift-x keymap
will toggle the rendering of LaTex in the associated preview view. Note
that the LaTex rendering will toggle even if the preview view is not the
current active item of its parent pane.