Unsuccessful attempt to implement double buffer
This was an interesting idea, to implement a double buffer for rendering
LaTex with MathJax in the previews. I found that if the preview page
becomes densely populated with LaTex, then processing can take a second
or a few. That in itself isn't so bad, however the physical hight of the
rendered LaTex compared to the hight of the page when MathJaxs hasn't
yet rendered it can be greatly different. The effect is that the scroll
position of the preview (that is usually maintained with normal
markdown) can't be maintained. Let me try to explain with a scenario.

I have a long page with alot of LaTex that is rendered in the preview,
and I have positioned my scroll bar st I am at the very bottom of the
preview. I edit my source, and the preview trys to update. Because there
is a large amount of rendering to be done it initially pops in with just
the non equation content, and the page height is dramatically reduced,
lets say half the size of it with equations. The physicall scroll
location (in pixels) cannot be maintained, because that location is
twice as long as what is now being displayed, so the physicall scroll is
halfed (even the actual scroll bar remains at the bottom of its slider).
The LaTex finally pops in, and the page height is doubled. The sliders
physical scroll position can be maintained, but that is only half the
height of the new page, and so the position of the slider is now half
way along, and we are no longer looking at the end of the preview.

Maybe that made sense, maybe it didn't :D Bascially this was an attempt
to render the LaTex in a backbuffer, and then copy that content over to
the preview window. However I realise that actually my implementation
doesn't actually work. If I have to copy and paste (essentially) html
blocks, then of course the actual images, svg's etc loaded to that need
to be rerendered. They will probs still be in some cache, but the point
is its still not quick enough or good enough, and the copy and paste
process essentially has the same problem as loading the processed html
directly into the preview.

There are some interesting things I learned, such as using callbacks
when MathJax's has finished processing some DOM elements, see
render.coffee (37) for that, and other things you can diff, so I think
its worth keeping the code for learning purposes. But you shouldn't
continue development on this branch :D

Couple of ideas however.

1. Configure the preview to maintain a constant scroll percentage, not a
   scroll based on some descript number of pixels. That would fix part
   of the problem, and maybe not to hard to implement.

2. You actually need to have two equal priority buffers, that you can
   change the visibility (css) of. There is no way to achive what I want
   by moving DOM elements around, I'll just end up with another instance
   of the probelm

So lets see what can be done ;D