Typeset LaTeX before html hits the preview window
Previously, MathJax would process LaTeX for the preview window only
after the html of the preview window had been updated with the output
from the markdown parser. This mean that you would see the equations
pop in, which is not a great user experience.

Now MathJax processes the output of the markdown parser before it is
pushed to the preview window. A much cleaner UX.