Remove backtick quotes at beg and end of inline eq
In certain scenarios (a know one is when an inline eq starts a paragraph
block) the backtick quotes inserted in the preprocessor will NOT be
parsed to a '<code>' block, leaving backtick quotes in the final LaTeX
render. Right now I simply assume that if they are present in the parsed
markdown, then that is a mistake and remove them.

Note that I use a 'gm' modifier on my string, which right now means if a
backtick quote is found at EOL for a multiline inline eq, that will also
be removed.