Fix backtick quotes at beg and end of inline eq
The backticks that were appearing at the beg and end of certain inline
eq's actually pointed to a deeper problem. I originally thought it was
simply because a paragraph block started with <script> tags (after
having been pushed through the preprocessor), and <script> tags
according to the markdown syntax should be block level elements.

If <script> tags are embedded within (and not placed at the begining of)
a paragraph block, then the markdown parse will treat them as plain
text, not html tags (hence the `` -> <code> tunneling trick). However
what I did not realise when I applied 6791b0e is that if a pragraph
block starts with an opening <script> tag, the entire block is treated
as a <script> and NOT identified as a paragraph, even though there may
be no closing </script> tag at the end of the block (I would guess that
marked simply looks for the first blank line). The consequence is that
NO markdown will be processed in that block.

This corrects this bug, by removing the removal code for the backtick
quotes, and simply surrounding the <script> blocks of inline equations
with a <span> set, which are inline not block level elements, so the
paragraph is now correctly identified.