Refactor LaTeX toggle and remove global config var
The toggling of LaTeX rendering was being facilitated by a global config
variable. This added an unnecessary layer to the command chain to toggle
LaTeX rendering. The global config var was also exposed in the user
settings, which was never an intention.

The refactored code removes the global config var and
MarkdownPreviewView's now directly observe the
'markdown-preview:toggle-render-latex' command.