Copy old MathJax output when < 2 eqs have changed
This experimental branch is designed to increase performance in the use
case that either 0 or 1 equation has changed. In these cases it is more
efficient to copy the old MathJax's output from the old html of the
preview window to the new html.

It is twitchy though, and there were unforseen issues, but read the
comments in mathjax-helper for details. They all boil down to am I
creating (perceptible) decreases in performance outside of this use
case, I certainly shouldn't be. I think my really unthought is that I
havn't been able to properly quantify the performance of this to the
main mathjax branch, so that should be the next step.